% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{as.data.frame.summary.bootstrap}
\alias{as.data.frame.summary.bootstrap}
\title{Produce a dataframe of bootstrap summarization}
\usage{
as.data.frame.summary.bootstrap(x, ...)
}
\description{
Produce a dataframe of bootstrap summarization
}
\author{
kholub
}
