% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_testing.R
\name{as.data.frame.summary.mcrt}
\alias{as.data.frame.summary.mcrt}
\title{Access Monte Carlo randomization test summary data}
\usage{
as.data.frame.summary.mcrt(x, ...)
}
\description{
Access Monte Carlo randomization test summary data
}
\author{
kholub
}
