% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{as.data.frame.summary.prm_power}
\alias{as.data.frame.summary.prm_power}
\title{Access a permutation test power analysis summarization}
\usage{
as.data.frame.summary.prm_power(x, ...)
}
\arguments{
\item{x}{a prm_power object}
}
\description{
Access a permutation test power analysis summarization
}
\examples{
gen_data <- function(){ data.frame(x = c(rnorm(50), rnorm(50, 1)), class = rep(c('a', 'b'), each = 50)) }
mcrp <- mcr_power(gen_data, c(mean = function(df){ mean(df[df$class == 'a', 'x']) - mean(df[df$class == 'b', 'x']) }, 
                              median = function(df){ median(df[df$class == 'a', 'x']) - median(df[df$class == 'b', 'x']) }),
    'class')
s <- summary(mcrp, .05, 'two.sided')
as.data.frame(s)

}
\author{
kholub
}
