% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_tests.R
\name{as.data.frame.summary.prmt}
\alias{as.data.frame.summary.prmt}
\title{Access permutation test summary data}
\usage{
as.data.frame.summary.prmt(x, ...)
}
\description{
Access permutation test summary data
}
\author{
kholub
}
