% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Perform boostrap resampling}
\usage{
bootstrap(data, ..., trials = 100)
}
\arguments{
\item{data}{data frame or matrix containing observations to be bootstrapped}

\item{...}{statistics to compute}

\item{trials}{number of resamplings to perform}
}
\description{
Perform bootstrap resampling for arbitrary statistics using tidy semantics
Note that the global mc.cores can be used to control level of parallelism in computing statistics on resamples.
}
\examples{
boot_results <- bootstrap(iris, sepal_ratio = sum(Sepal.Length) / sum(Sepal.Width),
                                petal_ratio = sum(Petal.Length) / sum(Petal.Width))

}
\author{
kholub
}
