% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization_tests.R
\name{confint.mcrt}
\alias{confint.mcrt}
\title{Compute empirical confidence intervals for statistics under Monte Carlo randomization testing}
\usage{
confint.mcrt(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class mcrt}

\item{parm}{a vector of statistic names to produce intervals for}

\item{level}{the width of confidence intervals}
}
\description{
Compute empirical confidence intervals for statistics under Monte Carlo randomization testing
}
\examples{
data(mpg)
sample1 <- mpg \%>\% filter(class == 'suv')
sample2 <- mpg \%>\% filter(class == 'compact')
mcr_results <- mcrd_test(sample1, sample2, median = median(hwy),
                                           mean = mean(hwy))
confint(mcr_results, parm = c("mean", "median"))

}
\author{
kholub
}
