% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{mcr_power}
\alias{mcr_power}
\title{Perform a power analysis of a randomization test}
\usage{
mcr_power(generate_data, statistics, group_var, alternative,
  statistic_names = names(statistics), test_trials = 100,
  power_trials = 100)
}
\description{
Perform a power analysis of a randomization test
}
\author{
kholub
}
