% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization_tests.R
\name{mcr_test}
\alias{mcr_test}
\title{Perform a Monte Carlo randomization test}
\usage{
mcr_test(data, statistics, group_var, statistic_names = names(statistics),
  trials = 1000)
}
\arguments{
\item{data}{a data frame to draw statistics from}

\item{statistics}{a function or vector of functions accepting a data frame, returning a scalar statistic}

\item{group_var}{a character, singular column name in the data representing groups.}

\item{statistic_names}{the names of statistics that are computed. must be specified if the supplied statistics do not have names}

\item{trials}{the number of Monte Carlo trials to perform}
}
\description{
Simulate a permutation test of arbitrary statistics using a simple randomization (i.e. non-clustered) scheme
Note that \code{\link{mcrd_test}} and \code{\link{mcr_anova}} convenience functions cover common use cases and should be preferred for interactive, non-programmatic use
If the "parallel" pacakge is installed, use the global mc.cores option to control level of parallelism
}
\examples{
library(dplyr)
data(mpg)
test_data <- mpg \%>\% filter(class \%in\% c('suv', 'compact'))
mcrd_res <- mcrd_test(test_data, c(median_difference = function(df) {
                                                                      group_medians <- df \%>\%
                                                                        group_by(class) \%>\%
                                                                        summarize(
                                                                          median = median(hwy))
                                                                       as.numeric(group_medians[group_medians$class == 'suv', 'median']) -
                                                                         as.numeric(group_medians[group_medians$class == 'compact', 'median'])
                                                                     })
                      )

}
\author{
kholub
}
