% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization_tests.R
\name{mcrd_test}
\alias{mcrd_test}
\title{Perform a two sample Monte Carlo randomization test for difference in arbitrary statistics}
\usage{
mcrd_test(sample1, sample2, ..., trials = 1000)
}
\arguments{
\item{sample1}{a data.frame representing observations from a single sample or treatment}

\item{sample2}{''}

\item{...}{named statistics to be computed on the}

\item{trials}{the number of Monte Carlo trials to perform}
}
\value{
a Monte Carlo Randomization Test object containing the results of the test
}
\description{
Randomization testing for assessing significance of difference in arbitrary test statistics across two samples.
If the "parallel" pacakge is installed, use the global mc.cores option to control level of parallelism
}
\examples{
data(mpg)
sample1 <- mpg \%>\% filter(class == 'suv')
sample2 <- mpg \%>\% filter(class == 'compact')
mcrd_res <- mcrd_test(sample1, sample2, mean = mean(hwy),
                                        median = median(hwy))

}
\author{
kholub
}
