% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_testing.R
\name{plot.mcrt}
\alias{plot.mcrt}
\title{Visualize the distribution of statistic comparisons between two samples}
\usage{
plot.mcrt(x, ..., nrows = 1, ncols = NULL)
}
\arguments{
\item{nrows}{the number of rows in the resulting graphic (each entry corresponds to a single statistic)}

\item{ncols}{the number of columns in the resulting graphic}

\item{object}{an mcrt object}
}
\description{
Visualize the distribution of statistic comparisons between two samples
}
\examples{
data(mpg)
sample1 <- mpg \%>\% filter(class == 'suv')
sample2 <- mpg \%>\% filter(class == 'compact')
mcr_results <- mcr_test(sample1, sample2, median = median(hwy),
                                          mean = mean(hwy))
plot(mcr_results)

}
\author{
kholub
}
