% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_tests.R
\name{plot.prmt}
\alias{plot.prmt}
\title{Visualize observed statistics against null distributions}
\usage{
plot.prmt(x, ..., nrows = 1, ncols = NULL)
}
\arguments{
\item{nrows}{the number of rows in the resulting graphic (each entry corresponds to a single statistic)}

\item{ncols}{the number of columns in the resulting graphic}

\item{object}{an prmt object}
}
\description{
Visualize observed statistics against null distributions
}
\examples{
data(mpg)
sample1 <- mpg \%>\% filter(class == 'suv')
sample2 <- mpg \%>\% filter(class == 'compact')
mcr_results <- mcrd_test(sample1, sample2, median = median(hwy),
                                          mean = mean(hwy))
plot(mcr_results)

}
\author{
kholub
}
