% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.summary.bootstrap}
\alias{print.summary.bootstrap}
\title{Print a bootstrap result summary}
\usage{
print.summary.bootstrap(x, ...)
}
\arguments{
\item{x}{an object of class summary.boostrap to be printed}
}
\description{
Print a bootstrap result summary
}
\author{
kholub
}
