% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_testing.R
\name{print.summary.mcrt}
\alias{print.summary.mcrt}
\title{Print a Monte Carlo randomization test summary}
\usage{
print.summary.mcrt(x, ...)
}
\description{
Print a Monte Carlo randomization test summary
}
\author{
kholub
}
