% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{print.summary.prm_power}
\alias{print.summary.prm_power}
\title{Print a power analysis summary of a permutation test}
\usage{
print.summary.prm_power(x, ...)
}
\description{
Print a power analysis summary of a permutation test
}
\author{
kholub
}
