% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_tests.R
\name{prm_test}
\alias{prm_test}
\title{Perform a full permutation test}
\usage{
prm_test(data, statistics, group_var, statistic_names = names(statistics),
  max_observations = 10)
}
\arguments{
\item{data}{a data frame to draw statistics from}

\item{statistics}{a function or vector of functions accepting a data frame, returning a scalar statistic}

\item{group_var}{a character, singular column name in the data representing groups.}

\item{statistic_names}{the names of statistics that are computed. must be specified if the supplied statistics do not have names}

\item{trials}{the number of Monte Carlo trials to perform}
}
\description{
This function should only be used for academic purposes or small sample sizes (time grows factorially with number of observations)
If the "parallel" pacakge is installed, use the global mc.cores option to control level of parallelism
}
\examples{
library(dplyr)
data(mpg)
test_data <- mpg \%>\% 
  filter(class \%in\% c('suv', 'compact')) \%>\%
  sample_n(7) # this just for computational feasibility of illustrative example
prm_res <- prm_test(test_data, c(median_difference = function(df) {
                                                                     group_medians <- df \%>\%
                                                                       group_by(class) \%>\%
                                                                       summarize(
                                                                         median = median(hwy))
                                                                     as.numeric(group_medians[group_medians$class == 'suv', 'median']) -
                                                                     as.numeric(group_medians[group_medians$class == 'compact', 'median'])
                                                                  }), 
                    'class')

}
\author{
kholub
}
