% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{summary.bootstrap}
\alias{summary.bootstrap}
\title{Summarize a bootstrap result}
\usage{
summary.bootstrap(object, level = 0.95, ...)
}
\arguments{
\item{object}{a bootstrap result to be summarized}

\item{empirical_ci_level}{the lower quantile used to compute empirical intervals on the bootstrapped statistic}
}
\description{
Summarize a bootstrap result
}
\examples{
boot_results <- bootstrap(iris, sepal_ratio = sum(Sepal.Length) / sum(Sepal.Width),
                                petal_ratio = sum(Petal.Length) / sum(Petal.Width))
summary(boot_results)

}
\author{
kholub
}
