% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{summary.prm_power}
\alias{summary.prm_power}
\title{Summarize a power analysis of a randomization test}
\usage{
summary.prm_power(object, alpha, alternative, ...)
}
\description{
Summarize a power analysis of a randomization test
}
\author{
kholub
}
