% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_tests.R
\name{summary.prmt}
\alias{summary.prmt}
\title{Summarize the results of a permutation test}
\usage{
summary.prmt(object, level = 0.05, alternative = "two.sided", ...)
}
\arguments{
\item{object}{a prmt object to be summarized}

\item{level}{the width of empirical confidence intervals}

\item{alternative}{the alternative used for tests of significance. one of 'less', 'greater', 'two.sided'}
}
\description{
Yields p values & CIs against null (empirical) distribution of statistics.
For two sided hypotheses, reported p-value is the minimum one sided p-value with a Bonferroni correction.
}
\examples{
data(mpg)
sample1 <- mpg \%>\% filter(class == 'suv')
sample2 <- mpg \%>\% filter(class == 'compact')
mcr_results <- mcrd_test(sample1, sample2, mean = mean(hwy))
summary(mcr_results)

}
\author{
kholub
}
