% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{as.data.frame.summary.prm_power}
\alias{as.data.frame.summary.prm_power}
\title{Access a permutation test power analysis summarization}
\usage{
as.data.frame.summary.prm_power(x, ...)
}
\description{
Access a permutation test power analysis summarization
}
\author{
kholub
}
