% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{confint.bootstrap}
\alias{confint.bootstrap}
\title{Compute confidence intervals for a bootstrapped statistic}
\usage{
confint.bootstrap(object, parm, level = 0.95, ...)
}
\description{
Compute confidence intervals for a bootstrapped statistic
}
