% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_testing.R
\name{confint.mcrt}
\alias{confint.mcrt}
\title{Produce confidence intervals for statistics under Monte Carlo randomization testing}
\usage{
confint.mcrt(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class mcrt}

\item{parm}{a vector of statistic names to produce intervals for}

\item{level}{the width of confidence intervals}
}
\description{
Produce confidence intervals for statistics under Monte Carlo randomization testing
}
\examples{
data(mpg)
sample1 <- mpg \%>\% filter(class == 'suv')
sample2 <- mpg \%>\% filter(class == 'compact')
mcr_results <- mcr_test(~ hwy, sample1, sample2, median = median(hwy))
confint(mcr_results, parm = "median")

}
\author{
kholub
}
