% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{mcrd_power}
\alias{mcrd_power}
\title{Perform a power analysis of a randomization test
Note this is a convinience wrapper for \code{\link{mcr_power}} in the case of two sample differences}
\usage{
mcrd_power(generate_sample1, generate_sample2, alternative, ...,
  statistic_names = names(statistics), test_trials = 100,
  power_trials = 100)
}
\description{
Perform a power analysis of a randomization test
Note this is a convinience wrapper for \code{\link{mcr_power}} in the case of two sample differences
}
\author{
kholub
}
