% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{plot.prm_power}
\alias{plot.prm_power}
\title{Visualize the power analysis of a randomization test
If alpha is left unspecified, the distribution of p-values is visualized. If alpha is specified, power is plotted as a function of alpha}
\usage{
plot.prm_power(x, ..., statistic, alternative, alpha = seq(0.01, 0.99, by =
  0.01))
}
\description{
Visualize the power analysis of a randomization test
If alpha is left unspecified, the distribution of p-values is visualized. If alpha is specified, power is plotted as a function of alpha
}
\author{
kholub
}
